Enhanced

+++ src/Controller/VideosController.php
@@ -2,12 +2,12 @@
 
 namespace App\Controller;
 
+use App\Controller\Trait\CheckTrait;
 use App\Entity\Videos;
 use App\Entity\Figures;
 use App\Form\VideoForm;
 use App\Form\AddVideoForm;
 use App\Service\Parameters;
-use App\Repository\VideosRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,7 +18,10 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class VideosController extends AbstractController
 {
 
+    use CheckTrait;
 
+
+    public function __construct(public Parameters $parameters){}
     #[Route('/ajout-video/{id}', name:'add_video')]
     #[IsGranted('ROLE_USER', message:"Veuillez confirmer votre compte")]
     /**
@@ -31,31 +34,35 @@ class VideosController extends AbstractController
      * @param Parameters $parameters
      * @return Response
      */
-    public function addVideo(Request $request, EntityManagerInterface $entityManager, int $id, VideosRepository $videoRepo, Parameters $parameters) :Response
+    public function addVideo(Request $request, EntityManagerInterface $entityManager, int $id) :Response
     {
         $figure = $entityManager->getRepository(Figures::class)->find($id);
+        $numberOfVideos = count($figure->getVideos());
+        //Avoid exceeding the maximum number of videos allowed.
+        if ($numberOfVideos >= $this->parameters::MAX_VIDEOS) {
+            $this->addFlash('warning', $this->parameters->getMessages('errors', ['max_reach' => 'videos']));
+            return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
+        }
         $form = $this->createForm(AddVideoForm::class, $figure);
         $form->handleRequest($request);
         if ($form->isSubmitted() === true && $form->isValid() === true) {
             $videos = $form->get('videos')->getData();
             if ($videos) {
                 foreach ($videos as $value) {
-                    // The maximum number of videos allowed.
-                    if ($videoRepo->countVideos($id)[1] >= $parameters::MAX_VIDEOS) {
-                        $this->addFlash('warning', $parameters->getMessages('error', ['max_reach' => 'videos']));
+                    //The maximum number of videos allowed.
+                    if ($numberOfVideos >= $this->parameters::MAX_VIDEOS) {
+                        $this->addFlash('warning', $this->parameters->getMessages('errors', ['max_reach' => 'videos']));
                         return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
                     }
 
-                    if (count($figure->getVideos()) > 0) {
-                        foreach ($figure->getVideos() as $src) {
-                            if ($src->getSrc() === $value->getSrc()) {
-                                $this->addFlash('danger', $parameters->getMessages('error', ['videos' => 'used']));
-                                return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
-                            }
-
+                    //Avoid duplicate videos.
+                    if ($numberOfVideos > 0) {
+                        if ($this->check($figure->getVideos(), $value->getSrc(), 'src') === true) {
+                            $this->addFlash('danger', $this->parameters->getMessages('errors', ['videos' => 'used']));
+                            return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
                         }
-                    }
 
+                    }
                     $embed = new Videos;
                     $embed->setSrc($value->getSrc());
                     $figure->addVideos($embed);
@@ -63,7 +70,7 @@ class VideosController extends AbstractController
                     $entityManager->flush();
                 }
 
-                $this->addFlash('success', $parameters->getMessages('feedback', ['success' => 'videos']));
+                $this->addFlash('success', $this->parameters->getMessages('feedback', ['success' => 'videos']));
                 return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
 
             }
@@ -85,7 +92,7 @@ class VideosController extends AbstractController
      * @param integer $video_id
      * @return Response
      */
-    public function editVideo(Request $request, EntityManagerInterface $entityManager, int $id, int $video_id, Parameters $parameters) :Response
+    public function editVideo(Request $request, EntityManagerInterface $entityManager, int $id, int $video_id) :Response
     {
         $figure = $entityManager->getRepository(Figures::class)->find($id);
         $video = $entityManager->getRepository(Videos::class)->find($video_id);
@@ -94,8 +101,9 @@ class VideosController extends AbstractController
         if ($form->isSubmitted() === true && $form->isValid() === true) {
             if (count($figure->getVideos()) > 0) {
                 foreach ($figure->getVideos() as $src) {
+                     //Avoid duplicate
                     if ($src->getSrc() === $form->get('src')->getData() && $src->getId() !== $video->getId()) {
-                        $this->addFlash('danger', $parameters->getMessages('error', ['videos' => 'used']));
+                        $this->addFlash('danger', $this->parameters->getMessages('errors', ['videos' => 'used']));
                         return $this->redirectToRoute('figuresdetails', ["slug" => $figure->getSlug()]);
                     }
 
@@ -105,7 +113,7 @@ class VideosController extends AbstractController
             $video->setSrc($form->get('src')->getData());
             $entityManager->persist($video);
             $entityManager->flush();
-            $this->addFlash('success','modifé avec succès');
+            $this->addFlash('success',$this->parameters->getMessages('feedback', ['edit' => 'message']));
             return $this->redirectToRoute('figuresdetails', ['slug' => $figure->getSlug()]);
         }
 
@@ -116,21 +124,27 @@ class VideosController extends AbstractController
 
     #[Route('/suppression-video/{id}/{video_id}',name:'delete_video')]
     #[IsGranted('ROLE_USER', message:"Veuillez confirmer votre compte")]
-    public function deleteVideo(EntityManagerInterface $entityManager, int $id , int $video_id, Parameters $parameters) :Response
+    public function deleteVideo(EntityManagerInterface $entityManager, int $id , int $video_id, Request $request) :Response
     {
         $figure = $entityManager->getRepository(Figures::class)->find($id);
         $videos = $entityManager->getRepository(Videos::class)->find($video_id);
-        if(!$videos){
-            $this->addFlash('danger', "Cette vidéo n'existe pas");
-            return $this->redirectToRoute('home');
+        $submittedToken = $request->request->get('token');
+        if ($this->isCsrfTokenValid('delete-item', $submittedToken)) {
+            if (!$videos) {
+                $this->addFlash('danger', ['unknown' => 'video']);
+                return $this->redirectToRoute('figuresdetails', ['slug' => $figure->getSlug()]);
+            }
+
+            $figure->removeVideos($videos);
+            $entityManager->remove($videos);
+            $entityManager->persist($figure);
+            $entityManager->flush();
+            $this->addFlash('success', $this->parameters->getMessages('feedback', ['delete' => 'message']));
+            return $this->redirectToRoute('figuresdetails', ['slug' => $figure->getSlug()]);
         }
 
-        $figure->removeVideos($videos);
-        $entityManager->remove($videos);
-        $entityManager->persist($figure);
-        $entityManager->flush();
-        $this->addFlash('success', $parameters->getMessages('feedback', ['delete' => 'message']));
-        return $this->redirectToRoute('home');
+        $this->addFlash('warning', $this->parameters->getMessages('errors', ['authenticate' => 'access']));
+        return $this->redirectToRoute('figuresdetails', ['slug' => $figure->getSlug()]);
 
     }
